chore: Split pgn info processing to own file
