+++ lib/toPgn.js
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-const { pgns, getField } = require('./fromPgn')
+const { getField } = require('./fromPgn')
+const { pgns } = require('./pgns')
 const _ = require('lodash')
 const BitStream = require('bit-buffer').BitStream
 const Int64LE = require('int64-buffer').Int64LE
